requirements.txt
requests==2.25.1
Flask==2.0.1
PyYAML==5.4.1
pytest==6.2.4
pytest-cov==2.12.1
vcrpy==4.1.1
PyGithub==1.55
